accept 64-bit input, large range correction
accepts 64-bit uints as input to add_raw rather than 32-bit inputs. the
large range correction range is now calculated dynamically based on
the number of bits used by the current settings (log2m and register
size).

implementation is based on the java-hll implementation, blog posts from
AK, and HyperLogLog in Practice by Heule et al.

https://research.neustar.biz/2013/01/24/hyperloglog-googles-take-on-engineering-hll/

+++ src/main.rs
@@ -15,7 +15,7 @@ fn main() {
     println!("n,explicit,hll,error");
 
     for i in 0..samples {
-        let e: u32 = rng.gen();
+        let e: u64 = rng.gen();
         explicit.insert(e);
         approx.add_raw(e);
